requirements.txt
# Anthropic Claude API (CRITIQUE - MVP Phase 1)
anthropic>=0.18.0

# Limitless API
httpx>=0.25.0
aiohttp>=3.9.0

# Notion API
notion-client>=2.2.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0
